refactor: move refresh job types to enums
