+++ backend/app/api/v1/services/refresh_match_events.py
@@ -16,8 +16,7 @@ from app.constants.external_apis import (
     API_EVENT_TYPE_VAR,
     API_FOOTBALL_EVENTS_ENDPOINT,
 )
-from app.models.enums import EventType
-from app.models.refresh_job import JobName
+from app.models.enums import EventType, JobName
 from app.schemas.match_events import MatchEventRefreshRow
 from app.utils.refresh_summary import RefreshSummary
