+++ backend/app/models/enums.py
@@ -30,3 +30,17 @@ class EventType(str, enum.Enum):
     SUBSTITUTION = "substitution"
     VAR = "var"
     OTHER = "other"
+
+
+class JobName(str, enum.Enum):
+    STANDINGS_REFRESH = "standings_refresh"
+    MATCHES_REFRESH = "matches_refresh"
+    MATCH_EVENTS_REFRESH = "match_events_refresh"
+    TEAM_RANKINGS_REFRESH = "team_rankings_refresh"
+    PLAYER_STATS_REFRESH = "player_stats_refresh"
+
+
+class JobStatus(str, enum.Enum):
+    RUNNING = "running"
+    SUCCESS = "success"
+    FAILED = "failed"
