+++ backend/tests/integration/test_refresh_matches.py
@@ -2,9 +2,9 @@ from datetime import date, timedelta
 
 from app.api.v1.services import refresh_matches as refresh_matches_service
 from app.constants.external_apis import API_FOOTBALL_FIXTURES_ENDPOINT
-from app.models.enums import StageType, StatusType
+from app.models.enums import JobStatus, StageType, StatusType
 from app.models.match import Match
-from app.models.refresh_job import JobStatus, RefreshJob
+from app.models.refresh_job import RefreshJob
 from app.models.team import Team, TeamType
 from app.models.tournament import Tournament
 from app.models.tournament_team import TournamentTeam
